Fixes for clang compiler support.
Added 'TOrangeVector::classDescription()' specialization to
'DEFINE_TOrangeVector_classDescription' macro and removed its default implementation
in the template class (static 'st_classDescription' member must be initialized/specialized before
its use).

A similar problem arises in the included '.ppp' files where st_classDescription is used. So I
moved all '.ppp. includes after the "DEFINE_*"s.

Removed unnecessary/duplicate VWRAPPER/WRAPPER uses.

Added 'this->...' qualifier in template classes (that inherit a from a template class) for
accessing base class methods, otherwise clang does not find them (as per C++ standard).

Moved the TVarList and TAttributed*List definitions/implementations from 'orvector' to 'vars'.
Fixed includes in other files using the two headers accordingly.

Fixed some casts to int with more appropriate casts to size_t/uintptr_t.

Added explicit casts of constants to template class in include/stat.hpp
statistics functions.

(references #1219).

+++ source/orange/garbage.hpp
@@ -290,7 +290,7 @@ public:
 
   inline bool operator < (const GCPtr<T> &ps) const
   { return    (!counter && ps.counter)
-           || (int(counter->ptr) < int(ps.counter->ptr)); }
+           || (uintptr_t(counter->ptr) < uintptr_t(ps.counter->ptr)); }
 
 
   inline T *getUnwrappedPtr()
